tests: basic set of tests for analyzers
Given the most important logic is in analyzers, the tests are fousing mainly on
those. The `eval` package has been refactored by turning a loader into an interface
with very limited responsiblity. This allows introducing fakeloader to serve the data
while still covering as much of the logic as possible.